Rename to BaseControllr and move the controller and views to their own folder
