+++ app/views/candidate_interface/course_choices/replace_choices/base/choose_action.html.erb
