+++ app/views/candidate_interface/course_choices/replace_choices/base/confirm_choice.html.erb
