+++ app/views/candidate_interface/course_choices/replace_choices/base/pick_choice_to_replace.html.erb
