+++ app/views/candidate_interface/course_choices/replace_choices/base/replace_location.html.erb
