+++ config/routes.rb
@@ -200,15 +200,14 @@ Rails.application.routes.draw do
         get '/find-a-course' => 'course_choices/have_you_chosen#go_to_find', as: :go_to_find
         get '/find_a_course', to: redirect('/candidate/application/courses/find-a-course')
 
-        get '/replace' => 'course_choices/replace_choice#pick_choice_to_replace', as: :replace_course_choices
-        post '/replace' => 'course_choices/replace_choice#picked_choice'
-        get '/replace/:id' => 'course_choices/replace_choice#choose_action', as: :replace_course_choice
-        post '/replace/:id' => 'course_choices/replace_choice#route_action'
-        get '/replace/:id/location' => 'course_choices/replace_choice#replace_location', as: :replace_course_choice_location
-        post '/replace/:id/location' => 'course_choices/replace_choice#validate_location', as: :validate_new_course_choice_location
-
-        get '/replace/:id/confirm/:course_option_id' => 'course_choices/replace_choice#confirm_choice', as: :confirm_replacement_course_choice
-        get '/replace/:id/update/:course_option_id' => 'course_choices/replace_choice#update_choice', as: :update_replacement_course_choice
+        get '/replace' => 'course_choices/replace_choices/base#pick_choice_to_replace', as: :replace_course_choices
+        post '/replace' => 'course_choices/replace_choices/base#picked_choice'
+        get '/replace/:id' => 'course_choices/replace_choices/base#choose_action', as: :replace_course_choice
+        post '/replace/:id' => 'course_choices/replace_choices/base#route_action'
+        get '/replace/:id/location' => 'course_choices/replace_choices/base#replace_location', as: :replace_course_choice_location
+        post '/replace/:id/location' => 'course_choices/replace_choices/base#validate_location', as: :validate_new_course_choice_location
+        get '/replace/:id/confirm/:course_option_id' => 'course_choices/replace_choices/base#confirm_choice', as: :confirm_replacement_course_choice
+        get '/replace/:id/update/:course_option_id' => 'course_choices/replace_choices/base#update_choice', as: :update_replacement_course_choice
 
         get '/provider' => 'course_choices/provider_selection#new', as: :course_choices_provider
         post '/provider' => 'course_choices/provider_selection#create'
